requirements.txt
pytest>=9.0.2
pytest-html>=4.2.0
requests>=2.32.5
responses>=0.26.0
python-dotenv>=1.2.2